Move ScriptFile to separate namespace: Reko.Core.Scripts
