+++ src/Core/ILoader.cs
@@ -18,8 +18,8 @@
  */
 #endregion
 
-using Reko.Core;
 using Reko.Core.Assemblers;
+using Reko.Core.Scripts;
 using System;
 using System.Collections.Generic;
 using System.Linq;
