+++ src/Core/Project.cs
@@ -18,9 +18,8 @@
  */
 #endregion
 
-using Reko.Core;
 using Reko.Core.Lib;
-using Reko.Core.Serialization;
+using Reko.Core.Scripts;
 using System;
 using System.Collections.Generic;
 using System.Collections.Specialized;
