+++ src/Core/Serialization/ProjectLoader.cs
@@ -18,8 +18,8 @@
  */
 #endregion
 
-using Reko.Core;
 using Reko.Core.Configuration;
+using Reko.Core.Scripts;
 using System;
 using System.Collections.Generic;
 using System.Linq;
