+++ src/Core/Serialization/ProjectSaver.cs
@@ -18,6 +18,7 @@
  */
 #endregion
 
+using Reko.Core.Scripts;
 using System;
 using System.Collections;
 using System.Collections.Generic;
