+++ src/Decompiler/Loading/Loader.cs
@@ -22,7 +22,7 @@ using Reko.Core;
 using Reko.Core.Assemblers;
 using Reko.Core.Configuration;
 using Reko.Core.Memory;
-using Reko.Core.Serialization;
+using Reko.Core.Scripts;
 using Reko.Core.Services;
 using System;
 using System.Collections.Generic;
